requirements.txt
asgiref==3.7.2
attrs==23.1.0
autopep8==2.0.4
black==23.12.1
click==8.1.7
coverage==7.3.4
Django==5.0
django-nose==1.4.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0
flake8==6.1.0
gunicorn==21.2.0
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
mccabe==0.7.0
mypy-extensions==1.0.0
nose==1.3.7
packaging==23.2
pathspec==0.12.1
pep8==1.7.1
platformdirs==4.1.0
pycodestyle==2.11.1
pyflakes==3.1.0
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
rpds-py==0.15.2
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
uritemplate==4.1.1
psycopg2>=2.7.5
